requirements.txt
openai
flask
load_dotenv